+++ cmd/cce/nodes_test.go
@@ -22,6 +22,7 @@ import (
 	"strings"
 
 	cce "github.com/smartedgemec/controller-ce"
+	"github.com/smartedgemec/controller-ce/swagger"
 	"github.com/smartedgemec/controller-ce/uuid"
 
 	. "github.com/onsi/ginkgo"
@@ -846,6 +847,409 @@ var _ = Describe("/nodes", func() {
 				"nodes_dns_configs",
 				"cannot delete node_id %s: record in use in nodes_dns_configs",
 			),
+		)
+	})
+
+	Describe("GET /nodes/{node_id}/interfaces", func() {
+		DescribeTable("200 OK",
+			func() {
+				clearGRPCTargetsTable()
+				nodeCfg := createAndRegisterNode()
+				By("Sending a GET /nodes/{node_id}/interfaces request")
+				resp, err := apiCli.Get(fmt.Sprintf("http://127.0.0.1:8080/nodes/%s/interfaces", nodeCfg.nodeID))
+				Expect(err).ToNot(HaveOccurred())
+				defer resp.Body.Close()
+
+				By("Verifying a 200 OK response")
+				Expect(resp.StatusCode).To(Equal(http.StatusOK))
+
+				By("Reading the response body")
+				body, err := ioutil.ReadAll(resp.Body)
+				Expect(err).ToNot(HaveOccurred())
+
+				var ifaces swagger.InterfaceList
+
+				By("Unmarshalling the response")
+				Expect(json.Unmarshal(body, &ifaces)).To(Succeed())
+
+				By("Verifying the created node was returned")
+				Expect(ifaces).To(Equal(
+					swagger.InterfaceList{
+						Interfaces: []swagger.InterfaceSummary{
+							{
+								ID:                "if0",
+								Description:       "interface0",
+								Driver:            "kernel",
+								Type:              "none",
+								MACAddress:        "mac0",
+								VLAN:              0,
+								Zones:             nil,
+								FallbackInterface: "",
+							},
+							{
+								ID:                "if1",
+								Description:       "interface1",
+								Driver:            "kernel",
+								Type:              "none",
+								MACAddress:        "mac1",
+								VLAN:              1,
+								Zones:             nil,
+								FallbackInterface: "",
+							},
+							{
+								ID:                "if2",
+								Description:       "interface2",
+								Driver:            "kernel",
+								Type:              "none",
+								MACAddress:        "mac2",
+								VLAN:              2,
+								Zones:             nil,
+								FallbackInterface: "",
+							},
+							{
+								ID:                "if3",
+								Description:       "interface3",
+								Driver:            "kernel",
+								Type:              "none",
+								MACAddress:        "mac3",
+								VLAN:              3,
+								Zones:             nil,
+								FallbackInterface: "",
+							},
+						},
+					},
+				))
+			},
+			Entry("GET /nodes/{node_id}/interfaces"),
+		)
+
+		DescribeTable("404 Not Found",
+			func() {
+				By("Sending a GET /nodes/{node_id}/interfaces/{interface_id} request")
+				clearGRPCTargetsTable()
+				nodeCfg := createAndRegisterNode()
+				resp, err := apiCli.Get(
+					fmt.Sprintf("http://127.0.0.1:8080/nodes/%s/interfaces/%s",
+						nodeCfg.nodeID,
+						uuid.New()))
+				Expect(err).ToNot(HaveOccurred())
+				defer resp.Body.Close()
+
+				By("Verifying a 404 Not Found response")
+				Expect(resp.StatusCode).To(Equal(http.StatusNotFound))
+			},
+			Entry("GET /nodes/{node_id}/interfaces/{interface_id} with nonexistent ID"),
+		)
+	})
+
+	Describe("PATCH /nodes/{node_id}/interfaces", func() {
+		var (
+			nodeCfg *nodeConfig
+		)
+
+		BeforeEach(func() {
+			clearGRPCTargetsTable()
+			nodeCfg = createAndRegisterNode()
+		})
+
+		DescribeTable("200 OK",
+			func(reqStr string, expectedNodeResp *cce.NodeResp) {
+				By("Sending a PATCH /nodes/{node_id}/interfaces request")
+				switch strings.Count(reqStr, "%s") {
+				case 1:
+					reqStr = fmt.Sprintf(reqStr, nodeCfg.nodeID)
+				case 5:
+					trafficPolicyID := postTrafficPolicies()
+					reqStr = fmt.Sprintf(reqStr, nodeCfg.nodeID, trafficPolicyID, trafficPolicyID, trafficPolicyID,
+						trafficPolicyID)
+				}
+				resp, err := apiCli.Patch(
+					fmt.Sprintf("http://127.0.0.1:8080/nodes/%s/interfaces", nodeCfg.nodeID),
+					"application/json",
+					strings.NewReader(reqStr))
+				Expect(err).ToNot(HaveOccurred())
+				defer resp.Body.Close()
+
+				By("Verifying a 200 OK response")
+				Expect(resp.StatusCode).To(Equal(http.StatusOK))
+
+				By("Getting the updated node")
+				updatedNodeResp := getNode(nodeCfg.nodeID)
+
+				By("Verifying the node was updated")
+				expectedNodeResp.SetID(nodeCfg.nodeID)
+				Expect(updatedNodeResp.NetworkInterfaces).To(Equal(expectedNodeResp.NetworkInterfaces))
+			},
+			Entry(
+				"PATCH /nodes/{node_id}/interfaces with network interfaces",
+				`
+					{
+						"interfaces": [
+							{
+								"id": "if0",
+								"description": "CHANGED IN PATCH",
+								"driver": "userspace",
+								"type": "upstream",
+								"mac_address": "mac0",
+								"vlan": 50,
+								"zones": null,
+								"fallback_interface": ""
+							},
+							{
+								"id": "if1",
+								"description": "interface1",
+								"driver": "kernel",
+								"type": "none",
+								"mac_address": "mac1",
+								"vlan": 1,
+								"zones": null,
+								"fallback_interface": ""
+							},
+							{
+								"id": "if2",
+								"description": "interface2",
+								"driver": "kernel",
+								"type": "none",
+								"mac_address": "mac2",
+								"vlan": 2,
+								"zones": null,
+								"fallback_interface": ""
+							},
+							{
+								"id": "if3",
+								"description": "interface3",
+								"driver": "kernel",
+								"type": "none",
+								"mac_address": "mac3",
+								"vlan": 3,
+								"zones": null,
+								"fallback_interface": ""
+							}
+						]
+					}
+				`,
+				&cce.NodeResp{
+					NetworkInterfaces: []*cce.NetworkInterface{
+						{
+							ID:                "if0",
+							Description:       "CHANGED IN PATCH",
+							Driver:            "userspace",
+							Type:              "upstream",
+							MACAddress:        "mac0",
+							VLAN:              50,
+							Zones:             nil,
+							FallbackInterface: "",
+						},
+						{
+							ID:                "if1",
+							Description:       "interface1",
+							Driver:            "kernel",
+							Type:              "none",
+							MACAddress:        "mac1",
+							VLAN:              1,
+							Zones:             nil,
+							FallbackInterface: "",
+						},
+						{
+							ID:                "if2",
+							Description:       "interface2",
+							Driver:            "kernel",
+							Type:              "none",
+							MACAddress:        "mac2",
+							VLAN:              2,
+							Zones:             nil,
+							FallbackInterface: "",
+						},
+						{
+							ID:                "if3",
+							Description:       "interface3",
+							Driver:            "kernel",
+							Type:              "none",
+							MACAddress:        "mac3",
+							VLAN:              3,
+							Zones:             nil,
+							FallbackInterface: "",
+						},
+					},
+				}),
+		)
+
+		DescribeTable("400 Bad Request",
+			func(reqStr string, expectedResp string) {
+				By("Sending a PATCH /nodes/{node_id}/interfaces request")
+				if strings.Contains(reqStr, "%s") {
+					reqStr = fmt.Sprintf(reqStr, nodeCfg.nodeID)
+				}
+				resp, err := apiCli.Patch(
+					fmt.Sprintf("http://127.0.0.1:8080/nodes/%s/interfaces", nodeCfg.nodeID),
+					"application/json",
+					strings.NewReader(reqStr))
+				Expect(err).ToNot(HaveOccurred())
+				defer resp.Body.Close()
+
+				By("Verifying a 400 Bad Request")
+				Expect(resp.StatusCode).To(Equal(http.StatusBadRequest))
+
+				By("Reading the response body")
+				body, err := ioutil.ReadAll(resp.Body)
+				Expect(err).ToNot(HaveOccurred())
+
+				By("Verifying the response body")
+				Expect(string(body)).To(Equal(expectedResp))
+			},
+			Entry(
+				"PATCH /nodes/{node_id}/interfaces by leaving the driver of an interface empty",
+				`
+				{
+					"interfaces": [
+						{
+							"id": "if0",
+							"description": "",
+							"driver": "",
+							"type": "",
+							"mac_address": "mac0",
+							"vlan": 50,
+							"zones": null,
+							"fallback_interface": ""
+						}
+					]
+				}
+				`,
+				"Validation failed: network_interfaces[0].driver must be one of [kernel, userspace]"),
+			Entry(
+				"PATCH /nodes/{node_id}/interfaces by leaving the type of an interface empty",
+				`
+					{
+						"interfaces": [
+							{
+								"id": "if0",
+								"description": "",
+								"driver": "kernel",
+								"type": "",
+								"mac_address": "mac0",
+								"vlan": 50,
+								"zones": null,
+								"fallback_interface": ""
+							}
+						]
+					}
+					`,
+				"Validation failed: network_interfaces[0].type must be one of [none, upstream, "+
+					"downstream, bidirectional, breakout]"),
+		)
+
+		DescribeTable("404 Not Found",
+			func(reqStr string, expectedResp string) {
+				By("Sending a PATCH /nodes/{node_id}/interfaces request")
+				resp, err := apiCli.Patch(
+					fmt.Sprintf("http://127.0.0.1:8080/nodes/%s/interfaces", nodeCfg.nodeID),
+					"application/json",
+					strings.NewReader(reqStr))
+				Expect(err).ToNot(HaveOccurred())
+				defer resp.Body.Close()
+
+				By("Verifying a 404 Not Found")
+				Expect(resp.StatusCode).To(Equal(http.StatusNotFound))
+
+				By("Reading the response body")
+				body, err := ioutil.ReadAll(resp.Body)
+				Expect(err).ToNot(HaveOccurred())
+
+				By("Verifying the response body")
+				Expect(string(body)).To(Equal(expectedResp))
+			},
+			Entry(
+				"PATCH /nodes with network interfaces",
+				`
+				{
+					"interfaces": [
+						{
+							"id": "if03",
+							"description": "interface0",
+							"driver": "userspace",
+							"type": "upstream",
+							"mac_address": "mac0",
+							"vlan": 50,
+							"zones": null,
+							"fallback_interface": ""
+						},
+						{
+							"id": "if1",
+							"description": "interface1",
+							"driver": "kernel",
+							"type": "none",
+							"mac_address": "mac1",
+							"vlan": 1,
+							"zones": null,
+							"fallback_interface": ""
+						},
+						{
+							"id": "if2",
+							"description": "interface2",
+							"driver": "kernel",
+							"type": "none",
+							"mac_address": "mac2",
+							"vlan": 2,
+							"zones": null,
+							"fallback_interface": ""
+						},
+						{
+							"id": "if3",
+							"description": "interface3",
+							"driver": "kernel",
+							"type": "none",
+							"mac_address": "mac3",
+							"vlan": 3,
+							"zones": null,
+							"fallback_interface": ""
+						}
+					]
+				}
+				`,
+				"Network Interface if03 not found"),
+		)
+
+		DescribeTable("500 Internal server error",
+			func(reqStr string, expectedResp string) {
+				By("Sending a PATCH /nodes request")
+				if strings.Contains(reqStr, "%s") {
+					reqStr = fmt.Sprintf(reqStr, nodeCfg.nodeID)
+				}
+				resp, err := apiCli.Patch(
+					fmt.Sprintf("http://127.0.0.1:8080/nodes/%s/interfaces", nodeCfg.nodeID),
+					"application/json",
+					strings.NewReader(reqStr))
+				Expect(err).ToNot(HaveOccurred())
+				defer resp.Body.Close()
+
+				By("Verifying a 500 Internal server error")
+				Expect(resp.StatusCode).To(Equal(http.StatusInternalServerError))
+
+				By("Reading the response body")
+				body, err := ioutil.ReadAll(resp.Body)
+				Expect(err).ToNot(HaveOccurred())
+
+				By("Verifying the response body")
+				Expect(string(body)).To(Equal(expectedResp))
+			},
+			Entry("PATCH /nodes/{node_id}/interfaces without all interfaces",
+				`
+				{
+					"interfaces": [
+						{
+							"id": "if0",
+							"description": "interface0",
+							"driver": "userspace",
+							"type": "upstream",
+							"mac_address": "mac0",
+							"vlan": 50,
+							"zones": null,
+							"fallback_interface": ""
+						}
+					]
+				}
+				`,
+				"error bulk updating network interfaces: rpc error: code = FailedPrecondition desc = Network Interface if1 missing from request"), //nolint:lll
 		)
 	})
 })